Key, Door & Light
There's a door in the scene with a light above it. By entering the collision trigger area of the door, the light above it will turn red for not having a key. After destroying all the targets in the scene, the key will appear. Then when the player enters the door's trigger area again, the light will be green and the door will open.