Added Radar
I wanted to make a UI element of a radar to pinpoint targets in the stage by following this tutorial: https://www.youtube.com/watch?v=MICA7pCPFO4

It works as intended with a bit of scuff. The tutorial did not account for if the actor is destroyed so every time while testing I destroyed an actor, I'd get thousands of error messages after I exit play. But hey! The game still runs!

All that's left for the level (for me at least) is just a few more targets.